pyproject.toml
[project]
name = "risingwave-test"
version = "0.1.0"
description = "Real-time E-commerce Conversion Funnel with RisingWave, dbt, and Iceberg"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka>=2.12.2",
    "dash>=3.4.0",
    "dash-bootstrap-components>=1.4.0",
    "plotly>=5.0.0",
    "pandas>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "dbt>=1.0.0.40.14",
    "dbt-core>=1.10.19",
    "dbt-risingwave>=1.10.0",
    "kafka-python>=2.3.0",
    "sqlalchemy>=2.0.46",
    "duckdb>=1.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "dagster>=1.10.0",
    "dagster-dbt>=0.26.0",
    "dagster-webserver>=1.10.0",
    "marimo>=0.10.0",
    "pyspark[connect]>=3.5.0,<4.0.0",
    "pyiceberg[pyarrow]>=0.8.0",
    "s3fs>=2024.1.0",
    "nbformat>=5.0.0",
    "aiohttp>=3.13.3",
    "psutil>=7.2.2",
]

[tool.dagster]
module_name = "orchestration.definitions"
project_name = "realtime_funnel"
